+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 ### Added
 - Add shortcuts instead of insert key
 
+### Changed
+- [#182] - Enable the first note-on when tone portamento is set before the playing start position (thanks [@freq-mod])
+
+[@freq-mod]: https://github.com/freq-mod
+
+[#182]: https://github.com/rerrahkr/BambooTracker/issues/182
+
 ## v0.3.5 (2020-02-15)
 ### Added
 - Support C86CTL